Slimmed down RankingList file to < 200 lines
